get_plan/requirements.txt
openai
langchain
langchain_core
langchain_openai
googlemaps
streamlit
python-dotenv
flask
flask-cors
requests
flask_mysqldb